fix(make): cast sample to f64 before multiply to support 64bit feature
i16::MAX was cast to f32, which fails to compile when rodio/64bit is
enabled and the sample type is f64. Cast through f64 explicitly so the
arithmetic is valid under both the default and all-decoders feature sets.

+++ src/sound_make.rs
@@ -150,7 +150,7 @@ fn generate_wav(frequency: f32, duration: Duration, amplify: f32) -> Result<Vec<
     let num_channels = source.channels().get();
 
     let samples: Vec<i16> = source
-        .map(|s| (s.clamp(-1.0, 1.0) * i16::MAX as f32).round() as i16)
+        .map(|s| ((s as f64).clamp(-1.0, 1.0) * i16::MAX as f64).round() as i16)
         .collect();
 
     let bits_per_sample = 16u16;
